+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.4.0 2019-05-17
+* add Content Security Policy (CSP) support
+* add Cross Origin Resource Sharing (CORS) support
+* remove transpile step (reduce dependencies)
+
 ## 2.1.0 2019-02-20
 * merge global, component-sepcific config into component object (overwrites default)